+++ protected/humhub/modules/user/models/fieldtype/BaseType.php
@@ -352,17 +352,14 @@ class BaseType extends Model
      *
      * @param User $user
      * @param bool $raw
-     * @return string
+     * @param bool $encode
+     * @return string|null
      */
-    public function getUserValue(User $user, $raw = true): ?string
+    public function getUserValue(User $user, bool $raw = true, bool $encode = true): ?string
     {
-        $internalName = $this->profileField->internal_name;
+        $value = $user->profile->{$this->profileField->internal_name} ?? '';
 
-        if ($raw) {
-            return $user->profile->$internalName;
-        } else {
-            return Html::encode($user->profile->$internalName);
-        }
+        return $encode ? Html::encode($value) : $value;
     }
 
     /**